requirements.txt
eventlet
flask
flask-socketio
build
geopy
gpxpy
jinja2
lxml
numpy
pytz
sqlalchemy
pyyaml
requests
scipy
timezonefinder
